refactor: extract shared config test helpers to eliminate duplication across 9 files
