+++ src/config-pr-feedback-commands.test.ts
@@ -1,13 +1,12 @@
 import { describe, it, expect } from "bun:test";
 import { writeFileSync, mkdirSync } from "fs";
-import { join, dirname } from "path";
-import { useTempDir } from "./test-utils.ts";
+import { join } from "path";
+import { useTempDir, makeConfigTestHelpers } from "./test-utils.ts";
 import {
   parseConfigFile,
   applyEnvOverrides,
   parseCLIArgs,
   resolveConfig,
-  getConfigFilePath,
   DEFAULTS,
 } from "./config.ts";
 
@@ -206,21 +205,7 @@ describe("parseCLIArgs — hooks.prFeedback", () => {
 
 describe("resolveConfig — hooks.prFeedback precedence", () => {
   const ctx = useTempDir();
-
-  function env(
-    extra?: Record<string, string>,
-  ): Record<string, string | undefined> {
-    return { RALPHAI_HOME: join(ctx.dir, "home"), ...extra };
-  }
-
-  function writeGlobalConfig(
-    cwd: string,
-    config: Record<string, unknown>,
-  ): void {
-    const filePath = getConfigFilePath(cwd, env());
-    mkdirSync(dirname(filePath), { recursive: true });
-    writeFileSync(filePath, JSON.stringify(config));
-  }
+  const { env, writeGlobalConfig } = makeConfigTestHelpers(ctx);
 
   it('returns default "" when no overrides', () => {
     const cwd = join(ctx.dir, "repo-prfc-default");
